+++ src/views/mainPage/components/buttonContainer.js
@@ -2,6 +2,7 @@ import { restaurants } from "../../../constants/restaurantData.js";
 import restaurantList from "./restaurantList.js";
 import { $ } from "../../../utils/domHelpers.js";
 import button from "../../../components/@common/button.js";
+import buttonHandler from "../../../eventHandler/button.js";
 
 const buttonContainer = () => {
   const buttonContainer = $(".button-container");
@@ -24,27 +25,7 @@ const buttonContainer = () => {
   const $addButton = $("#add-button");
   const $cancelButton = $("#cancel-button");
 
-  $addButton.addEventListener("click", (event) => {
-    event.preventDefault();
-
-    const $category = $("#category");
-    const $name = $("#name");
-    const $distance = $("#distance");
-    const $description = $("#description");
-    const $link = $("#link");
-
-    const newRestaurant = {
-      category: $category.value,
-      title: $name.value,
-      distance: $distance.value,
-      description: $description.value,
-      link: $link.value,
-    };
-
-    restaurants.push(newRestaurant);
-    //TODO: 모달 닫기 로직 추가
-    restaurantList();
-  });
+  $addButton.addEventListener("click", buttonHandler);
 };
 
 export default buttonContainer;
